[fs] revert "xfs: fix bogus space reservation in xfs_iomap_write_allocate"
Message-id: <20170606225456.18732-1-billodo@redhat.com>
Patchwork-id: 179353
O-Subject: [RHEL7.4 PATCH BZ 1446484] Revert "[fs] xfs: fix bogus space reservation in xfs_iomap_write_allocate"
Bugzilla: 1446484
RH-Acked-by: Brian Foster <bfoster@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

This reverts commit 7145a2b3e13d017799fbe20d48701df658acf687.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1446484
Brew: : https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=13357670
Tested: Locally

Commit 7145a2b3 (upstream 0af32fb46) was merged to remove a bogus space
reservation in xfs_iomap_write_allocate, stating "there is no reason to
reserve disk blocks in the transaction when allocating blocks for delalloc
space as we already reserved the space when creating the delalloc extent."
However, removal of the space reservation ultimately yielded an assert
in xfs_iomap_write_allocate(), manifested in debug kernels only:

...
[   79.695460] t_blk_res = 0
[   79.698102] XFS: Assertion failed: tp->t_blk_res_used <= tp->t_blk_res, file: fs/xfs/xfs_trans.c, line: 315
...
[   79.943382] Call Trace:
_blk_res_used <=[   79.945858]  [<ffffffffc0543778>] xfs_trans_mod_sb+0x2e8/0x320 [xfs]
 tp->t_blk_res, [   79.953639]  [<ffffffffc04c1c10>] xfs_alloc_ag_vextent+0x1a0/0x2c0 [xfs]
file: fs/xfs/xfs[   79.961776]  [<ffffffffc04c3238>] xfs_alloc_vextent+0x818/0xcb0 [xfs]
_trans.c, line: [   79.969651]  [<ffffffffc04da161>] xfs_bmap_btalloc+0x451/0x8f0 [xfs]
315
[   79.977438]  [<ffffffffc04da624>] xfs_bmap_alloc+0x24/0x40 [xfs]
[   79.983999]  [<ffffffffc04dbcd4>] xfs_bmapi_write+0x684/0xf70 [xfs]
[   79.990315]  [<ffffffffc052fc0b>] xfs_iomap_write_allocate+0x16b/0x370 [xfs]
[   79.997416]  [<ffffffffc0511167>] xfs_map_blocks+0x307/0x430 [xfs]
[   80.003642]  [<ffffffffc0511c40>] xfs_do_writepage+0x190/0x6c0 [xfs]
[   80.010032]  [<ffffffffb97d0b27>] write_cache_pages+0x287/0x630
[   80.015996]  [<ffffffffc0511ab0>] ? xfs_aops_discard_page+0x150/0x150 [xfs]
[   80.022999]  [<ffffffffb963c369>] ? sched_clock+0x9/0x10
[   80.028343]  [<ffffffffb96e8245>] ? local_clock+0x25/0x30
[   80.033775]  [<ffffffffb971c28f>] ? lock_release_holdtime.part.29+0xf/0x1a0
[   80.040785]  [<ffffffffc05115b5>] xfs_vm_writepages+0xc5/0x100 [xfs]
[   80.047178]  [<ffffffffb97d2391>] do_writepages+0x21/0x50
[   80.052608]  [<ffffffffb98857d0>] __writeback_single_inode+0x40/0x590
[   80.059082]  [<ffffffffb9886310>] writeback_sb_inodes+0x290/0x590
[   80.065211]  [<ffffffffb98866af>] __writeback_inodes_wb+0x9f/0xd0
[   80.071338]  [<ffffffffb9886a33>] wb_writeback+0x353/0x580
...

The assert condition hadn't occurred prior to commit 7145a2b3, as the
"bogus" space reservation was sufficient (nres=4, instead of zero) to
avoid the assert designed to occur when t_blk_res_used exceeded t_blk_res
in xfs_trans_mod_sb() for the XFS_TRANS_SB_FDBLOCKS case:
		if (delta < 0) {
			tp->t_blk_res_used += (uint)-delta;
			ASSERT(tp->t_blk_res_used <= tp->t_blk_res);
		}

The XFS_TRANS_SB_FDBLOCKS case is used for blocks that are actually
allocated by the transaction, and so must be reserved beforehand.
IOW, the case is normally run for blocks allocated by the transaction,
as opposed to delayed allocation conversion in the XFS_TRANS_SB_RES_FDBLOCKS
case. So here a block is being allocated for what was effectively a hole.
We're writing back a page that was a target of a previous write, a
condition that shouldn't occur during page writeback. For this case,
the page must be backed by physically allocated blocks or delalloc blocks.

In particular a write that is deemed as short results in an unmap and removing
the extent. After the lead-up condition in which a short write happens, the
extent is truncated away and the writeback tries to reallocate it, triggering
the assert. A new delalloc block isn't allocated and so one gets allocated at
writeback time, despite the fact that it should have accounted for it at
delalloc write time.

Although this revert fixes the assert issue, the root problem isn't
actually the transaction reservation. Rather, for some reason, allocation
is occuring when: a new block from a context where we expect to be doing
delalloc -> real extent conversion. Commit 7145a2b3 effectively uncovered
the secondary issue, and resulted in the assert. Removing this secondary
effect is indicated now, while developing a solution for the root cause
is ongoing. As far as upstream goes, the root cause is non-existent:
hch's iomap patchset in June 2016 removed the possibility of occurrence,
and the commit to take out the "bogus" space reservation (upstream
0af32fb46) was made later. Consequently, no upstream fix was ever required.

Signed-off-by: Bill O'Donnell <billodo@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>