[fs] pnfs: add a new mechanism to select a layout driver according to an ordered list
Message-id: <489f14a2cd0e3d8874cd30868e5b9b189fa8f0e0.1496929363.git.bcodding@redhat.com>
Patchwork-id: 179481
O-Subject: [RHEL7 fs PATCH 2/3] pnfs: add a new mechanism to select a layout driver according to an ordered list
Bugzilla: 1459899
RH-Acked-by: Scott Mayhew <smayhew@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

Currently, the layout driver selection code always chooses the first one
from the list. That's not really ideal however, as the server can send
the list of layout types in any order that it likes. It's up to the
client to select the best one for its needs.

This patch adds an ordered list of preferred driver types and has the
selection code sort the list of available layout drivers according to it.
Any unrecognized layout type is sorted to the end of the list.

For now, the order of preference is hardcoded, but it should be possible
to make this configurable in the future.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-by: J. Bruce Fields <bfields@fieldses.org>
Signed-off-by: Anna Schumaker <Anna.Schumaker@Netapp.com>
(cherry picked from commit ca440c383a588091cae9fbce610b86a6e9d961ad)
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1459899
Signed-off-by: Rafael Aquini <aquini@redhat.com>