[char] random: mix in architectural randomness earlier in extract_buf()
Message-id: <E1dJECt-0002XX-SX@gondobar>
Patchwork-id: 179504
O-Subject: [RHEL7.4 v2 PATCH 1/51] random: mix in architectural randomness earlier in extract_buf()
Bugzilla: 1270982 1457281
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL7.4 Bugzilla #1457281

commit 85a1f77716cf546d9b9c42e2848b5712f51ba1ee

Previously if CPU chip had a built-in random number generator (i.e.,
RDRAND on newer x86 chips), we mixed it in at the very end of
extract_buf() using an XOR operation.

We now mix it in right after the calculate a hash across the entire
pool.  This has the advantage that any contribution of entropy from
the CPU's HWRNG will get mixed back into the pool.  In addition, it
means that if the HWRNG has any defects (either accidentally or
maliciously introduced), this will be mitigated via the non-linear
transform of the SHA-1 hash function before we hand out generated
output.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Rafael Aquini <aquini@redhat.com>