[char] random: drop trickle mode
Message-id: <E1dJED1-0002Z8-PK@gondobar>
Patchwork-id: 179538
O-Subject: [RHEL7.4 v2 PATCH 8/51] random: drop trickle mode
Bugzilla: 1270982 1457281
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL7.4 Bugzilla #1457281

commit 95b709b6be49e4ff3933ef6a5b5e623de2713a71

The add_timer_randomness() used to drop into trickle mode when entropy
pool was estimated to be 87.5% full.  This was important when
add_timer_randomness() was used to sample interrupts.  It's not used
for this any more --- add_interrupt_randomness() now uses fast_mix()
instead.  By elimitating trickle mode, it allows us to fully utilize
entropy provided by add_input_randomness() and add_disk_randomness()
even when the input pool is above the old trickle threshold of 87.5%.

This helps to answer the criticism in [1] in their hypothetical
scenario where our entropy estimator was inaccurate, even though the
measurements in [2] seem to indicate that our entropy estimator given
real-life entropy collection is actually pretty good, albeit on the
conservative side (which was as it was designed).

[1] http://eprint.iacr.org/2013/338.pdf
[2] http://eprint.iacr.org/2012/251.pdf

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Rafael Aquini <aquini@redhat.com>