[char] random: simplify accounting logic
Message-id: <E1dJEDF-0002bq-3i@gondobar>
Patchwork-id: 179545
O-Subject: [RHEL7.4 v2 PATCH 20/51] random: simplify accounting logic
Bugzilla: 1270982 1457281
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL7.4 Bugzilla #1457281

commit ee1de406ba6eb1e01f143fe3351e70cc772cc63e

This logic is exactly equivalent to the old logic, but it should
be easier to see what it's doing.

The equivalence depends on one fact from outside this function:
when 'r->limit' is false, 'reserved' is zero.  (Well, two facts;
the other is that 'reserved' is never negative.)

Cc: Jiri Kosina <jkosina@suse.cz>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Signed-off-by: Greg Price <price@mit.edu>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Rafael Aquini <aquini@redhat.com>