[char] random: If we have arch_get_random_seed*(), try it before blocking
Message-id: <E1dJEDP-0002ds-9L@gondobar>
Patchwork-id: 179516
O-Subject: [RHEL7.4 v2 PATCH 29/51] random: If we have arch_get_random_seed*(), try it before blocking
Bugzilla: 1270982 1457281
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL7.4 Bugzilla #1457281

commit 331c6490c7f10dcf263712e313b7c0bc7fb6d77a

If we have arch_get_random_seed*(), try to use it for emergency refill
of the entropy pool before giving up and blocking on /dev/random.  It
may or may not work in the moment, but if it does work, it will give
the user better service than blocking will.

Reviewed-by: Ingo Molnar <mingo@kernel.org>
Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Rafael Aquini <aquini@redhat.com>