[char] random: remove unneeded hash of a portion of the entropy pool
Message-id: <E1dJEDU-0002f0-R3@gondobar>
Patchwork-id: 179529
O-Subject: [RHEL7.4 v2 PATCH 34/51] random: remove unneeded hash of a portion of the entropy pool
Bugzilla: 1270982 1457281
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL7.4 Bugzilla #1457281

commit 85608f8e16c28f818f6bb9918958d231afa8bec2

We previously extracted a portion of the entropy pool in
mix_pool_bytes() and hashed it in to avoid racing CPU's from returning
duplicate random values.  Now that we are using a spinlock to prevent
this from happening, this is no longer necessary.  So remove it, to
simplify the code a bit.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: George Spelvin <linux@horizon.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>