[char] random: clean up interrupt entropy accounting for archs w/o cycle counters
Message-id: <E1dJEDX-0002fS-3K@gondobar>
Patchwork-id: 179528
O-Subject: [RHEL7.4 v2 PATCH 36/51] random: clean up interrupt entropy accounting for archs w/o cycle counters
Bugzilla: 1270982 1457281
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL7.4 Bugzilla #1457281

commit 840f95077ffd640df9c74ad9796fa094a5c8075a

For architectures that don't have cycle counters, the algorithm for
deciding when to avoid giving entropy credit due to back-to-back timer
interrupts didn't make any sense, since we were checking every 64
interrupts.  Change it so that we only give an entropy credit if the
majority of the interrupts are not based on the timer.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: George Spelvin <linux@horizon.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>