[crypto] testmgr - Allow ecb(cipher_null) in FIPS mode
Message-id: <E1dLnR9-0001Vk-SK@gondobar>
Patchwork-id: 179831
O-Subject: [RHEL7.4 v5 PATCH 2/31] crypto: testmgr - Allow ecb(cipher_null) in FIPS mode
Bugzilla: 1459191
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL7.4 Bugzilla #1459191

commit 6175ca2ba7d2c4265d6d5590d699fbd907b24708

The cipher_null is not a real cipher, FIPS mode should not restrict its use.

It is used for several tests (for example in cryptsetup testsuite) and also
temporarily for reencryption of not yet encrypted device in cryptsetup-reencrypt tool.

Problem is easily reproducible with
  cryptsetup benchmark -c null

Signed-off-by: Milan Broz <gmazyland@gmail.com>
Acked-by: Stephan Mller <smueller@chronox.de>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Rafael Aquini <aquini@redhat.com>