[crypto] drbg - do not call drbg_instantiate in healt test
Message-id: <E1dLnRD-0001WQ-5h@gondobar>
Patchwork-id: 179836
O-Subject: [RHEL7.4 v5 PATCH 5/31] crypto: drbg - do not call drbg_instantiate in healt test
Bugzilla: 1459191
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Bruno E. O. Meneguele <brdeoliv@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL7.4 Bugzilla #1459191

commit d89a67134fcc7e863530624fe5a88dde0159cfb8

When calling the DRBG health test in FIPS mode, the Jitter RNG is not
yet present in the kernel crypto API which will cause the instantiation
to fail and thus the health test to fail.

As the health tests cover the enforcement of various thresholds, invoke
the functions that are supposed to enforce the thresholds directly.

This patch also saves precious seed.

Reported-by: Tapas Sarangi <TSarangi@trustwave.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Rafael Aquini <aquini@redhat.com>