[redhat] makefile: switch secureboot kernel image signing to release keys
Message-id: <21415fc26b5546a5c9b1a259af75f700999c9b0b.1497980251.git.aquini@redhat.com>
Patchwork-id: 179980
O-Subject: [RHEL7 PATCH 2/2] redhat: makefile: switch secureboot kernel image signing to release keys
Bugzilla:
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>

Since the next build should be RC we need to flip RELEASED_KERNEL to 1,
so that the build uses the production UEFI secureboot keys to sign kernel
image instead of the beta/devel one currently in use.

This is a RHEL7-only patch, not upstream.

Signed-off-by: Rafael Aquini <aquini@redhat.com>