[target] tcmu: prep for sync add/rm device support
Message-id: <1498591265-6293-2-git-send-email-mchristi@redhat.com>
Patchwork-id: 180212
O-Subject: [RHEL 7.4 PATCH 1/4] tcmu: prep for sync add/rm device support
Bugzilla: 1430225
RH-Acked-by: Chris Leech <cleech@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

- Patch description:

This patch adds the userspace header definitions needed to support
tcmu reconfigure. We are not supporting that operation in RHEL 7.4,
but the next patches are built on top of them upstream, so we
are adding the definitions, so our kernel/tools are compatibile
with upstream so in 7.5 we can easily port them.

- Differences from upstream:

This patch does not port everything from the above link. See above.

- Upstream status:

Not Merged. Patch posted upstream:

https://www.spinics.net/lists/target-devel/msg15710.html

- Brew:

http://brewweb.devel.redhat.com/brew/taskinfo?taskID=13517108

- BZ:

https://bugzilla.redhat.com/show_bug.cgi?id=1430225

- Testing:

Patches have been verified by the gluster and ceph team.

Signed-off-by: Mike Christie <mchristi@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>