[fs] locks: remove conditional lock release in middle of flock_lock_file
Message-id: <20170720131110.10595-1-jlayton@redhat.com>
Patchwork-id: 181012
O-Subject: [RHEL7.5 PATCH] locks: remove conditional lock release in middle of flock_lock_file
Bugzilla: 1461552
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>

As Linus pointed out:

    Say we have an existing flock, and now do a new one that conflicts. I
    see what looks like three separate bugs.

     - We go through the first loop, find a lock of another type, and
    delete it in preparation for replacing it

     - we *drop* the lock context spinlock.

     - BUG #1? So now there is no lock at all, and somebody can come in
    and see that unlocked state. Is that really valid?

     - another thread comes in while the first thread dropped the lock
    context lock, and wants to add its own lock. It doesn't see the
    deleted or pending locks, so it just adds it

     - the first thread gets the context spinlock again, and adds the lock
    that replaced the original

     - BUG #2? So now there are *two* locks on the thing, and the next
    time you do an unlock (or when you close the file), it will only
    remove/replace the first one.

...remove the "drop the spinlock" code in the middle of this function as
it has always been suspicious. This should eliminate the potential race
that can leave two locks for the same struct file on the list.

He also pointed out another thing as a bug -- namely that you
flock_lock_file removes the lock from the list unconditionally when
doing a lock upgrade, without knowing whether it'll be able to set the
new lock. Bruce pointed out that this is expected behavior and may help
prevent certain deadlock situations.

We may want to revisit that at some point, but it's probably best that
we do so in the context of a different patchset.

[ Upstream uses different spinlocks to protect the list, this version
  changes to using the i_lock ]

Reported-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
(cherry picked from commit 267f1128583074b575b90a58de4dcb12dd25af96)
RHBZ: https://bugzilla.redhat.com/show_bug.cgi?id=1461552
Signed-off-by: Jeff Layton <jlayton@redhat.com>
---

Note that this is only tested for compilation, but I think it's
sane enough to include this, given that we're early in the 7.5
cycle.

Signed-off-by: Rafael Aquini <aquini@redhat.com>