[fs] locks: require that flock->l_pid be set to 0 for file-private locks
Message-id: <20170724184710.20005-12-jlayton@redhat.com>
Patchwork-id: 181078
O-Subject: [RHEL7.5 PATCH 11/17] locks: require that flock->l_pid be set to 0 for file-private locks
Bugzilla: 1444778
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Neil Brown suggested potentially overloading the l_pid value as a "lock
context" field for file-private locks. While I don't think we will
probably want to do that here, it's probably a good idea to ensure that
in the future we could extend this API without breaking existing
callers.

Typically the l_pid value is ignored for incoming struct flock
arguments, serving mainly as a place to return the pid of the owner if
there is a conflicting lock. For file-private locks, require that it
currently be set to 0 and return EINVAL if it isn't. If we eventually
want to make a non-zero l_pid mean something, then this will help ensure
that we don't break legacy programs that are using file-private locks.

Cc: Neil Brown <neilb@suse.de>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
(cherry picked from commit 90478939dce096ed5b239cad16237dca0a59d66f)
BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1444778
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>