[fs] libceph: add an epoch_barrier field to struct ceph_osd_client
Message-id: <1501080593-8623-72-git-send-email-idryomov@redhat.com>
Patchwork-id: 181350
O-Subject: [RHEL7.5 fs PATCH 169/184] libceph: add an epoch_barrier field to struct ceph_osd_client
Bugzilla: 1450262
RH-Acked-by: Sage Weil <sweil@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

Cephfs can get cap update requests that contain a new epoch barrier in
them. When that happens we want to pause all OSD traffic until the right
map epoch arrives.

Add an epoch_barrier field to ceph_osd_client that is protected by the
osdc->lock rwsem. When the barrier is set, and the current OSD map
epoch is below that, pause the request target when submitting the
request or when revisiting it. Add a way for upper layers (cephfs)
to update the epoch_barrier as well.

If we get a new map, compare the new epoch against the barrier before
kicking requests and request another map if the map epoch is still lower
than the one we want.

If we get a map with a full pool, or at quota condition, then set the
barrier to the current epoch value.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-by: Ilya Dryomov <idryomov@gmail.com>
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit 58eb7932ae4d671d2a2377a1779eda96a2789b11)
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>