[fs] ceph: when seeing write errors on an inode, switch to sync writes
Message-id: <1501080593-8623-74-git-send-email-idryomov@redhat.com>
Patchwork-id: 181339
O-Subject: [RHEL7.5 fs PATCH 171/184] ceph: when seeing write errors on an inode, switch to sync writes
Bugzilla: 1450262
RH-Acked-by: Sage Weil <sweil@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

Currently, we don't have a real feedback mechanism in place for when we
start seeing buffered writeback errors. If writeback is failing, there
is nothing that prevents an application from continuing to dirty pages
that aren't being cleaned.

In the event that we're seeing write errors of any sort occur on an
inode, have the callback set a flag to force further writes to be
synchronous. When the next write succeeds, clear the flag to allow
buffered writeback to continue.

Since this is just a hint to the write submission mechanism, we only
take the i_ceph_lock when a lockless check shows that the flag needs to
be changed.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-by: "Yan, Zheng” <zyan@redhat.com>
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit 26544c623e741ac6445f8b1ae369ee32ae1794ad)
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	fs/ceph/file.c [ read/write interfaces in RHEL vs upstream ]
---