[fs] smb2: Enforce sec= mount option
Message-id: <1501827125-12680-2-git-send-email-lsahlber@redhat.com>
Patchwork-id: 181901
O-Subject: [RHEL7 PATCH fs 1/2] [bz 1083869] smb2: Enforce sec= mount option
Bugzilla: 1083869
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: Sachin Prabhu <sprabhu@redhat.com>

From: Sachin Prabhu <sprabhu@redhat.com>

Bugzilla: 1083869
brew: 13778479

If the security type specified using a mount option is not supported,
the SMB2 session setup code changes the security type to RawNTLMSSP. We
should instead fail the mount and return an error.

The patch changes the code for SMB2 to make it similar to the code used
for SMB1. Like in SMB1, we now use the global security flags to select
the security method to be used when no security method is specified and
to return an error when the requested auth method is not available.

For SMB2, we also use ntlmv2 as a synonym for nltmssp.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Acked-by: Pavel Shilovsky <pshilov@microsoft.com>
Acked-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit ef65aaede23f75977af56a8c330bb9be8c6e125c)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>