[tools] perf pmu: Support event aliases for non cpu// pmus
Message-id: <20170807084247.7531-17-jolsa@redhat.com>
Patchwork-id: 181963
O-Subject: [PATCH RHEL7.5 16/94] perf pmu: Support event aliases for non cpu// pmus
Bugzilla: 1478485
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1478485
https://bugzilla.redhat.com/show_bug.cgi?id=1478485

upstream
========
commit 231bb2aa32498cbebef1306889a02114e9dfc934
Author: Andi Kleen <ak@linux.intel.com>
Date: Fri Jan 27 18:03:39 2017 -0800

description
===========
The code for handling pmu aliases without specifying the PMU hardcoded
only supported the cpu PMU.

This patch extends it to work for all PMUs. We always duplicate the
event for all PMUs that have an matching alias.  This allows to
automatically expand an alias for all instances of a PMU (so for example
you can monitor all cache boxes with a single event)

Signed-off-by: Rafael Aquini <aquini@redhat.com>