[tools] perf tools: Use zfree() to avoid keeping dangling pointers
Message-id: <20170807084247.7531-25-jolsa@redhat.com>
Patchwork-id: 181964
O-Subject: [PATCH RHEL7.5 24/94] perf tools: Use zfree() to avoid keeping dangling pointers
Bugzilla: 1478485
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1478485
https://bugzilla.redhat.com/show_bug.cgi?id=1478485

upstream
========
commit 360e071b1822d40357e371357f24f6ec43f0a080
Author: Taeung Song <treeze.taeung@gmail.com>
Date: Wed Feb 1 21:34:07 2017 +0900

description
===========
The cases changed in this patch are for when we free but keep the
pointer to the freed area, which is not always a good idea.

Be more defensive and zero the pointer to avoid possible use after
free bugs to take more time to be detected.

[ rewrote commit log ]

RHEL7
=====
Conflicts with changes from upstream commit we did not take:
  2d055bf253c0 perf tools: Support setting different slots in a BPF map separately

Omitting ranges related hunks.

Signed-off-by: Rafael Aquini <aquini@redhat.com>